refactor(lotto): 디렉토리 구성
