+++ __tests__/CalculateTest.js
@@ -1,4 +1,4 @@
-import calculate from '../src/calculate.js';
+import calculate from '../src/utils/calculate.js';
 
 describe('calculate 기능 테스트', () => {
   test('구입금액을 수량으로 변환하는 기능 확인', () => {
