+++ __tests__/LottoMachineTest.js
@@ -1,4 +1,4 @@
-import lottoMachine from '../src/lottoMachine.js';
+import lottoMachine from '../src/domains/LottoDomain.js';
 
 describe('lottoMachine 테스트', () => {
   test('입력받은 숫자만큼 로또를 만드는 기능 확인', () => {
