+++ src/domains/LottoDomain.js
@@ -1,4 +1,4 @@
-import Lotto from './Lotto.js';
+import Lotto from '../Lotto.js';
 import { Random } from '@woowacourse/mission-utils';
 
 class LottoMachine {
