+++ src/views/prompt.js
